pyproject.toml
[tool.poetry]
name = "mcp_sever"
version = "0.1.0"
description = "Understainding and Creating MCP server"
package-mode = false
authors = ["Prasad Ovhal"]

[tool.poetry.dependencies]
python = "3.11.*"
torch = {version = "^2.5.1", source = "pytorch-cu121"}
torchvision = {version = "^0.20.1", source = "pytorch-cu121"}
torchaudio = {version = "^2.5.1", source = "pytorch-cu121"}
fastapi = "^0.135.1"
uvicorn = "^0.42.0"
pydantic = "^2.12.5"
redis = "^7.3.0"
sqlalchemy = "^2.0.48"
python-jose = "^3.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
slowapi = "^0.1.9"
httpx = "^0.28.1"
mcp = "^1.26.0"
langchain = "^1.2.12"
openai = "^2.29.0"
ollama = "^0.6.1"
chromadb = "^1.5.5"
sentence-transformers = "^5.3.0"
duckdb = "^1.5.0"
neo4j = "^6.1.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "*"

[[tool.poetry.source]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"